Move pFq to CMF from known_cmfs
