+++ ramanujantools/cmf/__init__.py
@@ -1,3 +1,5 @@
 from .cmf import CMF
+from .ffbar import FFbar
+from .pfq import pFq
 
-__all__ = ["CMF"]
+__all__ = ["CMF", "FFbar", "pFq"]
