--- ramanujantools/cmf/ffbar/__init__.py
@@ -1,5 +1,0 @@
-from .ffbar import FFbar
-
-__all__ = [
-    "FFbar",
-]
